Run ConDecon

Run ConDecon (continuous deconvolution) to estimate the cell abundances associated with each query bulk sample

Usage

RunConDecon(
  counts,
  latent,
  bulk,
  variable.features,
  max.iter = 5000,
  max.cent = 5,
  dims = 10,
  degree = 1,
  step = ifelse(max.iter <= 10000, max.iter, 10000),
  min.cent = 1,
  n = round(ncol(counts)/2),
  trainingset = NULL,
  sigma_min_cells = NULL,
  sigma_max_cells = NULL,
  verbose = FALSE,
  k = 1
)

Arguments

counts

single-cell counts matrix (features x cells)

latent

matrix of single-cell latent space (cells x dims)

bulk

matrix of query bulk data (features x samples)

variable.features

character vector of the most variable features

max.iter

size of the training dataset (default = 5,000)

max.cent

max number of centers in the Gaussian (default = 5)

dims

number of dimensions from latent (default = 10)

degree

degree of the polynomial used

step

manually parallelize building the training dataset

min.cent

min number of centers in the Gaussian (default = 1)

n

number of cells to be chosen to create the training dataset (default is half the number of cells in the counts matrix)

trainingset

pre-generated training dataset

sigma_min_cells

min number of cells that should be captured by the standard deviation of the Gaussian

sigma_max_cells

max number of cells that should be captured by the standard deviation of the Gaussian

verbose

logical indicating whether to print progress (default = FALSE)

k

Number of nearest neighbor cells aggregated together when calculating rank correlation

Value

ConDecon object with continuous deconvolution results

Examples

data(counts_gps)
data(latent_gps)
data(bulk_gps)
data(variable_genes_gps)

# For this example, we will reduce the training size to max.iter = 50 to reduce run time
ConDecon_obj = RunConDecon(counts = counts_gps, latent = latent_gps,
bulk = bulk_gps, variable.features = variable_genes_gps, max.iter = 50)